Handle the case when no `accept-language` header is passed
When no `accept-language` header is passed, the `Negotiator` returns `["*"]`, which is not a suitable value for the subsequent match function.

I addressed this by specifying the possible locales by passing them to the languages function. This way, when no language header is passed, the negotiator returns the passed languages as the available options for the match function to choose from.